Removed dependency to fenicstools Placed simulation scripts and files into simulation folder
